+++ mistralrs-quant/src/lib.rs
@@ -57,7 +57,7 @@ pub use lora::{
 };
 pub use unquantized::UnquantLinear;
 use utils::isq::apply_immediate_isq;
-pub use utils::{BitWiseOp, LeftshiftOp, NonZeroOp, UQFF_QUANT_TYPE_OFFSET};
+pub use utils::{log, BitWiseOp, LeftshiftOp, NonZeroOp, UQFF_QUANT_TYPE_OFFSET};
 
 use candle_nn::{Linear, Module};
 use serde::{Deserialize, Deserializer, Serialize};
